fix(types): make mypy strict — remove disable_error_code
Fix 122 surfaced findings with real narrowing; 1 centralized rowcount ignore.

+++ webservices/ratis_product_analyser/repositories/label_repository.py
@@ -97,7 +97,7 @@ def update_label_scan_result(
     scan: Scan,
     *,
     scanned_name: str,
-    price: Decimal,
+    price: int,
     product_ean: str | None,
     match_method: str | None,
     status: str,
